// src/types/menuTypes.ts
// React Imports
import type { ReactNode } from 'react'

// MUI Imports
import type { ChipProps } from '@mui/material/Chip'

// Type Imports
import type {
  SubMenuProps as VerticalSubMenuProps,
  MenuItemProps as VerticalMenuItemProps,
  MenuSectionProps as VerticalMenuSectionProps
} from '@menu/vertical-menu'
import type {
  SubMenuProps as HorizontalSubMenuProps,
  MenuItemProps as HorizontalMenuItemProps
} from '@menu/horizontal-menu'

// Vertical Menu Data
export type VerticalMenuItemDataType = Omit<VerticalMenuItemProps, 'children' | 'icon' | 'prefix' | 'suffix'> & {
  label: ReactNode
  icon?: string
  for: 'admin' | 'creator'
  options?: (VerticalMenuItemDataType | VerticalSubMenuDataType | VerticalSectionDataType)[]
  prefix?: ReactNode | ChipProps
  suffix?: ReactNode | ChipProps
}
export type VerticalSubMenuDataType = Omit<VerticalSubMenuProps, 'children' | 'icon' | 'prefix' | 'suffix'> & {
  children: VerticalMenuDataType[]
  icon?: string
  for: 'admin' | 'creator'
  options?: (VerticalMenuItemDataType | VerticalSubMenuDataType | VerticalSectionDataType)[]
  prefix?: ReactNode | ChipProps
  suffix?: ReactNode | ChipProps
  option?: any
}
export type VerticalSectionDataType = Omit<VerticalMenuSectionProps, 'children'> & {
  isSection: boolean
  children: VerticalMenuDataType[]
}
export type VerticalMenuDataType = VerticalMenuItemDataType | VerticalSubMenuDataType | VerticalSectionDataType

// Horizontal Menu Data
export type HorizontalMenuItemDataType = Omit<HorizontalMenuItemProps, 'children' | 'icon' | 'prefix' | 'suffix'> & {
  label: ReactNode
  icon?: string
  for: 'admin' | 'creator'
  options?: (VerticalMenuItemDataType | VerticalSubMenuDataType | VerticalSectionDataType)[]
  prefix?: ReactNode | ChipProps
  suffix?: ReactNode | ChipProps
}
export type HorizontalSubMenuDataType = Omit<HorizontalSubMenuProps, 'children' | 'icon' | 'prefix' | 'suffix'> & {
  children: HorizontalMenuDataType[]
  icon?: string
  for: 'admin' | 'creator'
  options?: (VerticalMenuItemDataType | VerticalSubMenuDataType | VerticalSectionDataType)[]
  prefix?: ReactNode | ChipProps
  suffix?: ReactNode | ChipProps
}
export type HorizontalMenuDataType = HorizontalMenuItemDataType | HorizontalSubMenuDataType